+++ src/viur_cli/deploy.py
@@ -2,7 +2,6 @@ import os
 import string
 import click
 import yaml
-import subprocess
 from .conf import config
 from . import cli, echo_error, echo_info, replace_vars
 from .update import create_req
@@ -138,67 +137,3 @@ def deploy(action, profile, ext, yes, additional_args):
 
         os.system(
             f'gcloud app deploy --project={conf["application_name"]} {" ".join(additional_args)} {yaml_file}')
-
-
-@cli.command(context_settings={"ignore_unknown_options": True})
-@click.argument("action", type=click.Choice(["backup"]))
-def enable(action):
-    """
-    Enable specific features for the project.
-
-    The 'enable' command allows enabling various features for the project. Currently, only the 'backup' action is
-    supported.
-
-    Args:
-        action (str): The action to perform. Currently, only 'backup' is supported.
-
-    Example:
-    ```bash
-    $ viur enable backup
-    ```
-
-    """
-    if action == "backup":
-        enable_gcp_backup()
-
-
-def enable_gcp_backup():
-    # Load the project Config
-    conf = config.get_profile("default")
-
-    # Create helper Variables
-    project_id = conf["application_name"]
-    bucket_name = f'backup-dot-{project_id}'
-    backup_bucket_command = f'gsutil mb -l EUROPE-WEST3 -p {project_id} gs://{bucket_name}'
-
-    # Create the Backup Bucket
-    try:
-        result = subprocess.run(
-            backup_bucket_command,
-            capture_output=True,
-            shell=True,
-        )
-        print(result)
-        if result.returncode != 0:
-            print('Error creating bucket.')
-
-    except Exception as e:
-        print(f'An Error Occured:\n {e} Please make sure you have the correct Google Cloud Access rights')
-
-    # Helper Variables for IAM
-    iam_roles = ["roles/storage.admin", "roles/datastore.importExportAdmin"]
-    service_worker_mail = f'{project_id}@appspot.gserviceaccount.com'
-
-    for r in iam_roles:
-        iam_roles_command = (f'gcloud projects add-iam-policy-binding {project_id} --member '
-                             f'serviceAccount:{service_worker_mail} --role {r}')
-        try:
-            subprocess.run(iam_roles_command, capture_output=True, shell=True)
-
-        except Exception as e:
-            print(f'An Error Occured during Roles {e}\n '
-                  f'Please make sure you have the correct Google Cloud Access rights'
-                  )
-            return
-
-    print('Success! It may take a while until you can use Gcloud Backups')
